split config with custom
